+++ CHANGELOG.md
@@ -14,6 +14,8 @@ params for s3 `list_objects` methods
 - Add SecurityHub service
 - Add Transfer service
 - Introducing `rusoto_signature`, a standalone crate for signing HTTP requests.
+- Make static credentials into a credential provider
+- Add anonymous credentials support
 
 ## [0.41.0] - 2019-10-07